refactor: restructure component architecture to match application hierarchy
- Flatten deeply nested component directories for better maintainability
- Organize components by feature/domain (scene/, shared/)
- Group scene components by specific pages (about/, chat/, contact/, projects/, resume/, secret/)
- Move reusable components to shared/ directory
- Update import paths to reflect new structure
- Maintain logical separation between UI concerns and business logic

This restructuring improves code organization, reduces import path complexity,
and aligns the file structure with the application's functional hierarchy.
