+++ src/components/Scene.tsx
@@ -1,22 +1,18 @@
 import { Canvas } from "@react-three/fiber";
 import { OrbitControls, OrthographicCamera, Sky, SoftShadows } from "@react-three/drei";
-import { CubeWithFaces } from "./CubeWithFaces";
-import { CameraController } from "./CameraController";
-import { BowlGroundPlane } from "./BowlGroundPlane";
+import { CubeWithFaces } from "./scene/CubeWithFaces";
+import { CameraController } from "./scene/CameraController";
+import { BowlGroundPlane } from "./scene/BowlGroundPlane";
 import { useResponsiveFaceSize } from "@/lib/hooks/useResponsiveFaceSize";
 import { Float } from "@react-three/drei";
 import { useAtomValue } from "jotai";
 import { isLoadedAtom, bgMotionAtom, cubeMotionAtom } from "@/atoms/atomStore";
-import { useState } from "react";
-
 
 export const Scene = () => {
     useResponsiveFaceSize();
     const isLoaded = useAtomValue(isLoadedAtom)
     const bgMotion = useAtomValue(bgMotionAtom);
     const cubeMotion = useAtomValue(cubeMotionAtom);
-    
-
 
     return (
         <div className={`canvas w-full h-full ${isLoaded ? "opacity-100" : "opacity-0"}`}>
