+++ src/components/scene/BowlGroundPlane.tsx
