+++ src/components/scene/chat/assistant-ui/tooltip-icon-button.tsx
