+++ src/components/shared/Face.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React from 'react';
 
 export const Face = ({ children }: { children: React.ReactNode }) => {
   return (
